Dont normalize mods! This is for ACA to do.
Unwanted "normalization" of modifiers in the library causes bugs. This is only supposed to happen in ACA, for generating a tolerance, it's a side-effect of using 3 modifiers the way we do, technically when generating these, the three modifiers are like 3 tolerances, so tolerance is mathematically not needed. Tolerance only exists so that users can supply one tolerance value whenever they dont use ACA, so all simpler cases.